Telegram Workspace/requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.3.3
attrs==18.2.0
autobahn==19.2.1
Automat==0.7.0
beautifulsoup4==4.7.1
bleach==3.1.0
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
click==7.1.1
colorama==0.4.3
constantly==15.1.0
cryptography==2.6.1
cssselect==1.1.0
dateparser==0.7.1
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
graphviz==0.13.2
hyperlink==18.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==17.5.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-core==4.6.1
keyboard==0.13.5
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
numpy==1.16.2
openpyxl==2.5.14
pandas==1.0.1
pandocfilters==1.4.2
parsel==1.5.2
praw==6.5.1
prawcore==1.0.1
Protego==0.1.16
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.5.2
PyHamcrest==1.9.0
pylint==2.4.4
pynput==1.4.5
pyOpenSSL==19.0.0
pyperclip==1.7.0
pyrsistent==0.15.7
python-binance==0.7.1
python-dateutil==2.8.0
python-telegram-bot==12.6.1
pytz==2018.9
pywin32==227
queuelib==1.5.0
regex==2019.2.21
requests==2.21.0
scipy==1.2.1
Scrapy==1.8.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.7.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.4.1
tzlocal==1.5.1
update-checker==0.16
urllib3==1.24.1
utils==1.0.0
virtualenv==20.0.15
w3lib==1.21.0
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.11.2
zipp==2.2.0
zope.interface==4.6.0
